requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
idna==3.10
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
